Added 'Optional' text for description of Booking/Room Dialog, Updated account routes to disallow deletion of self/root and changed minimum access level to 25 for editing/deleting accounts

+++ src/common/actions/index.js
@@ -764,6 +764,14 @@ export const handleDeleteAccount = userId => {
 	return (dispatch, getState) => {
 		const defaultUser = getState().accounts.items.find(u => u.userId === userId)
 
+		if (userId === 'root') {
+			// Dispatch a 'Show Snackbar' action
+			dispatch(showSnackbar({
+				message: `Unable to remove account: ${defaultUser.username}`
+			}))
+			return
+		}
+
 		// Dispatch a 'Delete Account' action
 		dispatch(deleteAccount(userId))
